feat(cli): add asset selector to burn command (#2955)
* feat(cli): add asset selector to burn command

* fix(cli): Properly use the filtered balances response
